Refactor navigation
