+++ app/hotels/src/navigation/allHotels/MapHeaderButton.android.js
