+++ app/hotels/src/navigation/allHotels/MapHeaderButton.ios.js
