+++ app/hotels/src/navigation/allHotels/__tests__/LocationPickerScreen.test.js
